# api/docs/swagger.yaml
definitions:
  models.CreateLabel:
    properties:
      color:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
  models.CreateTask:
    properties:
      description:
        type: string
      due_date:
        type: string
      task_list_id:
        type: string
      title:
        type: string
    type: object
...
  models.Label:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
  models.Task:
    properties:
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      task_list_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateLabel:
    properties:
      color:
        type: string
      name:
        type: string
    type: object
  models.UpdateResponse:
    properties:
      id:
        type: string
      updated_at:
        type: string
    type: object
  models.UpdateTask:
    properties:
      description:
        type: string
      due_date:
        type: string
      task_list_id:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
  /label:
    post:
      consumes:
      - application/json
      description: create a new label
      parameters:
      - description: label
        in: body
        name: label
        schema:
          $ref: '#/definitions/models.CreateLabel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Label'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Creates a new label
      tags:
      - label
  /label/{id}:
    delete:
      consumes:
      - application/json
      description: delete label
      parameters:
      - description: label_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete label
      tags:
      - label
    get:
      consumes:
      - application/json
      description: get label by id
      parameters:
      - description: label_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Label'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get label by id
      tags:
      - label
    put:
      consumes:
      - application/json
      description: update label
      parameters:
      - description: label_id
        in: path
        name: id
        required: true
        type: string
      - description: label
        in: body
        name: label
        schema:
          $ref: '#/definitions/models.UpdateLabel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update label
      tags:
      - label
  /task:
    post:
      consumes:
      - application/json
      description: create a new task
      parameters:
      - description: task
        in: body
        name: task
        schema:
          $ref: '#/definitions/models.CreateTask'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Creates a new task
      tags:
      - task
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: delete task
      parameters:
      - description: task_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete task
      tags:
      - task
    get:
      consumes:
      - application/json
      description: get task by id
      parameters:
      - description: task_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get task by id
      tags:
      - task
    put:
      consumes:
      - application/json
      description: update task
      parameters:
      - description: task_id
        in: path
        name: id
        required: true
        type: string
      - description: task
        in: body
        name: task
        schema:
          $ref: '#/definitions/models.UpdateTask'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UpdateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update task
      tags:
      - task
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"